requirements.txt
torch
torchvision
torchaudio
opencv-python
matplotlib
open3d
pillow
requests
timm
